pyproject.toml
[project]
name = "fdua-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.45.2",
    "azure-ai-documentintelligence>=1.0.0",
    "faiss-cpu>=1.9.0.post1",
    "fugashi>=1.4.0",
    "google-genai>=1.0.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.3",
    "mecab-python3>=1.0.10",
    "openai>=1.60.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.5",
    "pillow-heif>=0.21.0",
    "polars>=1.20.0",
    "pyarrow>=19.0.0",
    "pymupdf>=1.25.2",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "ragatouille>=0.0.8.post4",
    "rank-bm25>=0.2.2",
    "rank-llm>=0.20.3",
    "sudachidict-full>=20250129",
    "sudachipy>=0.6.10",
    "tqdm>=4.67.1",
    "unidic-lite>=1.0.8",
    "unstructured[md]>=0.14.8",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]
